+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Bug fixes & Enhancements:
 - [#312] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/312) Unable to apply server_profile_template (Invalid character '.)
+- [#360] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/360) Server Profile Template / Server Profile : iLO Settings --> Cannot set iLO Hostname
 
 # [v6.2.0-13]